+++ src/index.tsx
@@ -3,11 +3,11 @@ import CustomShaderMaterial from './vanilla'
 import { iCSMProps } from './types'
 
 export default forwardRef<CustomShaderMaterial, iCSMProps>(
-  ({ baseMaterial, fragmentShader, vertexShader, uniforms, ...rest }, ref) => {
+  ({ baseMaterial, fragmentShader, vertexShader, uniforms, cacheKey, ...rest }, ref) => {
     const material = useMemo(
-      () => new CustomShaderMaterial(baseMaterial, fragmentShader, vertexShader, uniforms),
-      [baseMaterial, fragmentShader, vertexShader, uniforms]
+      () => new CustomShaderMaterial({ baseMaterial, fragmentShader, vertexShader, uniforms, cacheKey }),
+      [baseMaterial, fragmentShader, vertexShader, uniforms, cacheKey]
     )
-    return <primitive dispose={undefined} object={material} ref={ref} attach="material" {...rest} />
+    return <primitive object={material} ref={ref} attach="material" {...rest} />
   }
 )